- Adding Auth - Missing DTOs - Removing innecessary models - Adding Profiles

+++ TeacherControl/TeacherControl.Core/Models/Course/CourseUserCredit.cs
@@ -9,7 +9,7 @@ namespace TeacherControl.Core.Models
     {
         //Add a status here for the user credits?
         public int Id { get; set; }
-        public float Credits { get; set; }
+        public double Credits { get; set; }
         public int StudentId { get; set; }
         public virtual User Student { get; set; }
         public int CourseId { get; set; }
